# docs/openapi/flowsense-openapi.yaml
openapi: 3.0.3
info:
  title: FlowSense API
  description: >
    REST API for FlowSense, an indoor/outdoor campus wayfinding system for AUF.
    Serves three clients (Kiosk, Mobile Companion, Admin Dashboard) via a
    versioned, feature-based, role-aware API. Public endpoints (map, navigation,
    search, sessions) require no authentication; administrative endpoints
    (annotation, hardware, assets, analytics, users, settings) require an
    authenticated admin session. ESP32 sensors never call this REST API —
    they publish exclusively over MQTT.
  version: "1.0.0"
servers:
  - url: /api/v1
    description: Versioned API root
...
security:
  - AdminSessionCookie: []
paths:
...
  /map/areas/{id}/floors:
    get:
      tags: [Map]
      summary: List floors for area
      description: Retrieves all floors associated with a specific building/area.
      security: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: List of floors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FloorListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /search:
    get:
      tags: [Search]
      summary: Search destinations
      description: Executes a query to find rooms, aliases, or personnel matching the user's input.
      security: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: kiosk_session_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResultsResponse"
...
  /hardware/devices/{id}:
    get:
      tags: [Hardware]
      summary: Get device
      description: Retrieves specific identity, assignment, and connection details for a device.
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Device detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Hardware]
      summary: Update device
      description: Modifies device metadata such as Device Name, mapped Node, or Zone.
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceUpdateInput"
      responses:
        "200":
          description: Device updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Hardware]
      summary: Decommission device
      description: Soft-deletes or decommissions a piece of hardware from active operations.
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          $ref: "#/components/responses/SuccessNoData"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: flowsense_admin_session
      description: >
        Long-lived secure HTTP cookie established after passwordless
        administrator authentication. Required for all Admin API endpoints.
    KioskDeviceToken:
      type: http
      scheme: bearer
      description: >
        Short-lived kiosk access token obtained via device credential
        exchange. Used to distinguish an authenticated Kiosk Client from
        an ordinary browser on public endpoints.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    SuccessNoData:
      description: Operation completed successfully
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          nullable: true
        message:
          type: string
          nullable: true
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: ROOM_NOT_FOUND
            message:
              type: string
              example: Destination does not exist.
...
    RoomType:
      type: string
      enum: [room, office, laboratory, facility, service, restroom, hall, other]
...
    DeviceType:
      type: string
      enum: [kiosk, sensor]
    DeviceStatus:
      type: string
      enum: [unregistered, registered, online, offline, disabled, decommissioned]
...
    GeoJSONGeometry:
      type: object
      description: GeoJSON geometry expressed in EPSG:3857, returned/accepted by the API as GeoJSON regardless of internal PostGIS storage.
      properties:
        type:
          type: string
          enum: [Point, LineString, Polygon, MultiPolygon, GeometryCollection]
        coordinates:
          type: array
          items: {}

    # ---------------- Campus: Area ----------------
...
    Floor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        area_id:
          type: integer
          format: int64
        glb_node_name:
          type: string
          nullable: true
        floor_order:
          type: integer
        elevation:
          type: number
          format: double
          nullable: true
        display_name:
          type: string
          nullable: true
          description: Additive (step 14, QA-93). What the kiosk calls the floor, e.g. "First floor".
        short_name:
          type: string
          nullable: true
          description: Additive (step 14, QA-93). The kiosk's floor button, e.g. "1F".
        navigable:
          type: boolean
        visible_in_kiosk:
          type: boolean
        active:
          type: boolean
        geometry:
          $ref: "#/components/schemas/GeoJSONGeometry"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    FloorListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Floor"

    # ---------------- Campus: Room ----------------
    Room:
      type: object
      properties:
        id:
          type: integer
          format: int64
        floor_id:
          type: integer
          format: int64
        room_code:
          type: string
        room_alias:
          type: string
        description:
          type: string
          nullable: true
        room_type:
          $ref: "#/components/schemas/RoomType"
        is_searchable:
          type: boolean
        is_navigable:
          type: boolean
        is_active:
          type: boolean
        node_id:
          type: integer
          format: int64
          nullable: true
          description: >-
            The room's navigation node (where routes to it end), or null until
            the room is placed in Map Annotation. Returned by GET /map/rooms
            and in GET /search room results; the kiosk passes it to
            POST /navigation/routes as a destination.
        geometry:
          $ref: "#/components/schemas/GeoJSONGeometry"
        image_path:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Personnel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        full_name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        is_active:
          type: boolean
...
    SearchResult:
      type: object
      properties:
        result_type:
          type: string
          enum: [room, personnel]
        room:
          $ref: "#/components/schemas/Room"
        personnel:
          $ref: "#/components/schemas/Personnel"
    SearchResultsResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/SearchResult"

    # ---------------- Hardware ----------------
    Device:
      type: object
      properties:
        id:
          type: integer
          format: int64
        device_id:
          type: string
        device_type:
          $ref: "#/components/schemas/DeviceType"
        name:
          type: string
          nullable: true
        serial_number:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/DeviceStatus"
        enabled:
          type: boolean
        mac_address:
          type: string
          nullable: true
        ip_address:
          type: string
          nullable: true
        last_ping_at:
          type: string
          format: date-time
          nullable: true
        last_data_at:
          type: string
          format: date-time
          nullable: true
        uptime_seconds:
          type: integer
          format: int64
          nullable: true
        discovered_at:
          type: string
          format: date-time
        registered_at:
          type: string
          format: date-time
          nullable: true
    DeviceUpdateInput:
      type: object
      properties:
        name:
          type: string
        map_node_id:
          type: integer
          format: int64
        area_id:
          type: integer
          format: int64
        enabled:
          type: boolean
...
    DeviceResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Device"
...